backend/requirements.txt
opencv-python
face-recognition
numpy
pymongo
python-dotenv
cloudinary
redis
requests
flask
flask-cors